requirements.txt
onnxruntime>=1.16.0
numpy>=1.24.0
Pillow>=10.0.0
scikit-learn>=1.3.0
